Support for field conditions.
This is pretty ugly, as we need to switch within dynamic containers - trying to find the field that a condition refers to is not particularly simple. It works though...
The number of properties (and therefore constructor parameters) in FieldBase is now getting pretty large. We may want to extract that into a single object.