Cache container fields, albeit slightly hackily
This is the core of the optimization: we only convert each container's fields once, because they're not specific to a particular module address.
This effectively removes almost all the duplication we previously experienced. It would be nice not to have the cache within ContainerJson, but we can address that later.