Retain overlays on a per address basis while editing
This means if you change instrument group and change back again, we go back to the previous settings rather than resetting.
This information is only stored while editing, which matches the behaviour on the TD-17 module for Multifx. (For instruments, it gets reset entirely each time - I suspect our behaviour is preferable now.)

The behaviour of overlays around cymbal sizes still needs work: when changing instrument, we'll need to force the sizing.