Handle data changes synchronously, rather than dispatching
This may be a short term measure: if we ever want to react to module changes on the fly, we'll need to revisit this. But for now, we only make changes in the data within the UI thread, and it's a lot easier to reason about the various updates this way.