Use sliders for numeric fields
Currently this is optional, but with the non-slider code is still available. We could make this a UI option if we really wanted.
It's nice and simple while you can only edit via the slider (but still see the results as text in a label). If we wanted a "slider + textbox" option, that gets trickier in terms of avoiding ping-pong of events.