Release 1.0.0-alpha05
This is primarily to address issue 97, where multiple MIDI ports are
detected, but only one is for a Roland drum module, so we should be
able to proceed.

There are some known schema issues in the TD-27 schema which are
addressed in the new codebase but not here (as keeping track of
things in both places is really tricky).

This is likely to be the last release of the "old" code, as the
"new" code is approaching feature parity.